requirements.txt
alembic==1.4.2
aniso8601==6.0.0
appdirs==1.4.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
autopep8==1.5.4
awscli==1.18.159
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
botocore==1.18.18
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.8
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==2.0.0
docformatter==1.3.1
docutils==0.15.2
ecdsa==0.14.1
email-validator==1.1.1
entrypoints==0.3
env==0.1.0
filelock==3.0.12
Flask==1.1.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
ipaddr==2.2.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jose==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.0
ldap3==2.4.1
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==4.2.0
msgpack==0.6.2

nbformat==5.0.7
nest-asyncio==1.4.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.1
passlib==1.7.2
pbr==5.5.0
pep517==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2020.8.13
pluggy==0.13.1
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.4
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycups==1.9.73
Pygments==2.7.1
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.17.3
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
pytoml==0.1.21
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.18
sqlparse==0.2.4
sshtunnel==0.1.4
systemd-python==234
terminado==0.9.1
testpath==0.4.4
testresources==2.0.1
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4


untokenize==0.1.1
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.5.4
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.1
WTForms==2.3.3
yapf==0.30.0
zipp==1.0.0
zope.interface==5.1.0
